+++ samples/sample-donut-chart-with-legend.html
@@ -2,41 +2,119 @@
 
 <head>
   <script src="/smartChartsNXT/public/smartChartsNXT.bundle.js" type="text/javascript"></script>
+
   <script>
     window.onload = function () {
       SmartChartsNXT.ready(function () {
-        let donutChart = new SmartChartsNXT.Chart({
-          "type":"DonutChart",
-          "title": "Browser Statistics and Trends",
+        let DonutChart = new SmartChartsNXT.Chart({
+          "type": "DonutChart",
+          "title": "Browser Market Share Worldwide ",
+          "subtitle": "As of Q3, 2017",
+          "minRadius": 30,
+          "maxRadius": 100,
+          "titleStyle": {
+            "fillColor":"#717171",
+            "borderColor":"none",
+            "fontFamily": "Lato",
+            "maxFontSize": "25"
+          },
+          "subtitleStyle": {
+            "fillColor":"#717171",
+            "borderColor":"none",
+            "fontFamily": "Lato",
+            "maxFontSize": "18"
+          },
           "outline": 2,
-          "canvasBorder": false,
-          "subTitle": "As of Q1, 2016",
+          "canvasBorder": true,
           "targetElem": "chartContainer",
           "bgColor": "white",
-          "showLegend": true,
+          "radialGradient": [{offset:50, opacity:-0.1},{offset:55, opacity:0},{offset:67, opacity:0.2}, {offset:83, opacity:0.2},{offset:95, opacity:0}], //[optional]
+          "legends":{
+            "enable" : true,
+            "top": 200,
+            "left": 10, 
+            "alignment": "horizontal",
+            "float": "bottom",
+            "color": "#000",
+            "bgColor": "#eee",
+            "hoverColor":"#999",
+            "fontSize": 14, 
+            "fontFamily": "Lato", 
+            "borderColor": "none",
+            "borderWidth": 3,
+            "borderOpacity": 1,
+            "opacity": 0.9,
+            "toggleType": false,
+            // "onLegendClick": function() {console.log("onClick-",this)},
+            // "onLegendHover": function() {console.log("onHover-",this)},
+            // "onLegendLeave": function() {console.log("onLeave-",this)}
+          },
+          "animated": true,
+          "tooltip": {
+            "content": function() {
+              return '<table>' +
+              '<tr><td><b>'+this.label+'</b> has global usage </td></tr>' +
+              '<tr><td> of <b>'+this.value+'% </b>Worldwide.</td></tr>' +
+              '</table>';
+            }
+          },
           "dataSet": [{
               "label": "Chrome",
-              "value": "72.6",
+              "value": "54.53",
               "color": "#F44336"
             },
             {
               "label": "IE",
-              "value": "5.7"
+              "value": "3.7",
+              "color": "#4682b4"
             },
             {
               "label": "Safari",
-              "value": "3.6"
+              "value": "14.61"
             },
             {
               "label": "Firefox",
-              "value": "16.9",
-              "color": "#FFC107"
+              "value": "6.07",
+              "color": "#FFC107",
+              "slicedOut": false
             },
             {
               "label": "Opera",
-              "value": "1.2"
+              "value": "3.9"
+            },
+            {
+              "label": "UC Browser",
+              "value": "7.88",
+              "color": "#f5deb3"
+            },
+            {
+              "label": "Others",
+              "value": "9.29"
             }
-          ]
+          ] //,
+          // events: {
+          //   afterRender: function (e) {
+          //     console.log(e, "event after rendering complete");
+          //   },
+          //   onInit: function (e) {
+          //     console.log(e, "event onInit");
+          //   },
+          //   afterParseData: function (e) {
+          //     console.log(e, "event afterParseData");
+          //   },
+          //   beforeSave: function (e) {
+          //     console.log(e, "event beforeSave");
+          //   },
+          //   afterSave: function (e) {
+          //     console.log(e, "event afterSave");
+          //   },
+          //   beforePrint: function (e) {
+          //     console.log(e, "event beforePrint");
+          //   },
+          //   afterPrint: function (e) {
+          //     console.log(e, "event afterPrint");
+          //   }
+          // }
         });
       });
     };
@@ -45,7 +123,7 @@
 
 </head>
 
-<body>
+<body style="overflow: hidden;">
   <div>
     <div id='chartContainer' style="width:100%;height:100%;">
     </div>
