+++ samples/sample-pie-chart-2d.html
@@ -48,7 +48,7 @@
             // "onLegendHover": function() {console.log("onHover-",this)},
             // "onLegendLeave": function() {console.log("onLeave-",this)}
           },
-          "animated": true,
+          "animated": false,
           "tooltip": {
             "content": function() {
               return '<table>' +
