+++ src/charts/pieChart/sliceSet.js
@@ -1,25 +1,25 @@
 "use strict";
 
 /**
- * pieSet.js
+ * sliceSet.js
  * @createdOn: 01-Nov-2017
  * @author: SmartChartsNXT
- * @version: 1.1.0
- * @description:This is a parent class of pie slice class which create all the pie slices. 
+ * @version: 1.0.0
+ * @description:This is a parent class of slice class which create all the slices. 
  */
 
 import defaultConfig from "./../../settings/config";
 import Point from "./../../core/point";
 import {Component} from "./../../viewEngin/pview";
-import PieSlice from "./pieSlice"; 
+import Slice from "./slice"; 
 
-class PieSet extends Component {
+class SliceSet extends Component {
   constructor(props) {
     super(props);
     this.state = {
       startAngle:0,
       endAngle:0,
-      pieTextFontSize: defaultConfig.theme.fontSizeMedium,
+      textFontSize: defaultConfig.theme.fontSizeMedium,
       showSliceMarker: true,
       hideMarkerText: false, 
       animPercent: null
@@ -41,10 +41,10 @@ class PieSet extends Component {
 
   componentDidMount() {
     this.renderCounter++; 
-    let pieSetBBox = this.ref.node.getBBox(); 
-    if(pieSetBBox.width > this.props.pieAreaWidth) {
+    let sliceSetBBox = this.ref.node.getBBox(); 
+    if(sliceSetBBox.width > this.props.areaWidth) {
       switch(this.renderCounter) {
-        case 1: this.setState({pieTextFontSize: defaultConfig.theme.fontSizeSmall}); break; 
+        case 1: this.setState({textFontSize: defaultConfig.theme.fontSizeSmall}); break; 
         case 2: this.setState({hideMarkerText: true}); break;
         case 3: this.setState({showSliceMarker: false}); break; 
       }
@@ -60,25 +60,26 @@ class PieSet extends Component {
 
   render() {
     return (
-      <g class='pie-set'>
-        {this.createPieSlices()}
+      <g class='slice-set'>
+        {this.createSlices()}
       </g>
     );
   }
 
-  createPieSlices() {
+  createSlices() {
     return this.props.dataSet.map((data, i) =>{
       let percent = Math.min(this.state.animPercent || data.percent, data.percent);
       this.state.startAngle = this.state.endAngle; 
       this.state.endAngle += (percent * 3.6);
       return (
-        <PieSlice index={i} rootNodeId={this.props.rootNodeId}
+        <Slice index={i} rootNodeId={this.props.rootNodeId}
           toggleEnabled={this.props.dataSet.length > 0 ? true : false} cx ={this.props.cx} cy={this.props.cy} 
-          width={this.props.width} height={this.props.height} offsetWidth={this.props.offsetWidth} offsetHeight={this.props.offsetHeight}
+          width={this.props.width} height={this.props.height} innerWidth={this.props.innerWidth} innerHeight={this.props.innerHeight}
+          offsetWidth={this.props.offsetWidth} offsetHeight={this.props.offsetHeight}
           data={data} startAngle={this.state.startAngle} endAngle={this.state.endAngle} strokeColor={this.props.strokeColor} 
-          strokeWidth={this.props.strokeWidth} rotateChart={this.rotateChart.bind(this)} slicedOut={data.slicedOut} fontSize={this.state.pieTextFontSize}
-          updateTip={this.props.updateTip} hideTip={this.props.hideTip} onRef={ref => this.slices["s"+i] = ref} parentCtx={this.global} 
-          showSliceMarker={this.state.showSliceMarker} hideMarkerText={this.state.hideMarkerText}
+          strokeWidth={this.props.strokeWidth} rotateChart={this.rotateChart.bind(this)} slicedOut={data.slicedOut} fontSize={this.state.textFontSize}
+          gradient={this.props.gradient}  updateTip={this.props.updateTip} hideTip={this.props.hideTip} 
+          onRef={ref => this.slices["s"+i] = ref} parentCtx={this.global} showSliceMarker={this.state.showSliceMarker} hideMarkerText={this.state.hideMarkerText}
         />
       );
     });
@@ -98,4 +99,4 @@ class PieSet extends Component {
   
 }
 
-export default PieSet; 
+export default SliceSet;
